Modularize standalone performance tests
